fix error models
