+++ TipCatDotNet.Api/Data/Models/HospitalityFacility/ModelStatus.cs
@@ -1,8 +1,8 @@
-namespace TipCatDotNet.Api.Data.Models
+namespace TipCatDotNet.Api.Data.Models.HospitalityFacility
 {
     public enum ModelStatus: byte
     {
-        InActive = 0,
+        Inactive = 0,
         Active = 1,
         Deleted = 2
     }
