+++ TipCatDotNet.Api/Data/Models/HospitalityFacility/Owner.cs
@@ -2,32 +2,31 @@ using System;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace TipCatDotNet.Api.Data.Models
+namespace TipCatDotNet.Api.Data.Models.HospitalityFacility
 {
     [Table("accounts")]
-    public class Account
+    public class Owner
     {
         [Column("id")]
         public int Id { get; set; }
         
+        [Column("user_id")]
+        public int UserId { get; set; }
+        
+        [ForeignKey("UserId")]
+        public virtual User User { get; set; }
+        
         [Column("name")]
-        [StringLength(250)]
+        [StringLength(200)]
         public string Name { get; set; }
-        
-        [Column("address")]
-        public string Address { get; set; }
-        
-        [Column("commercial_name")]
-        [StringLength(250)]
-        public string CommercialName { get; set; }
+
+        [Column("last_name")]
+        [StringLength(200)]
+        public string LastName { get; set; }
         
         [Column("email")]
         [StringLength(200)]
         public string Email { get; set; }
-        
-        [Column("phone")]
-        [StringLength(20)]
-        public string Phone { get; set; }
         
         [Column("created")]
         public DateTime Created { get; set; }
