--- TipCatDotNet.Api/Data/Models/PermissionLevel.cs
@@ -1,9 +1,0 @@
-namespace TipCatDotNet.Api.Data.Models
-{
-    public enum PermissionLevel: byte
-    {
-        Owner = 1,
-        Manager = 2,
-        Employee = 3
-    }
-}
